# packages/api-types/schema/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Cleanway API",
    "description": "Phishing protection API. Checks domains for safety. Your browsing data lives on your device — we only see domain names.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/user/aggregates": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Submit Weekly Aggregate",
        "description": "Submit weekly aggregate numbers from device.\nCalled once per week by the extension/app.\nServer stores ONLY numbers — no URLs, no details.",
        "operationId": "submit_weekly_aggregate_api_v1_user_aggregates_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/WeeklyAggregateSubmit"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/percentile": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get Percentile",
        "description": "Get user's percentile ranking based on weekly blocks.\n\"Safer than X% of users\" — computed from all users' aggregates.",
        "operationId": "get_percentile_api_v1_user_percentile_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PercentileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/score": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Sync Score",
        "description": "Sync security score NUMBER from device.\nScore is CALCULATED on device. Server stores the number only.\nDetails/breakdown stay on device.",
        "operationId": "sync_score_api_v1_user_score_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScoreSync"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/device": {
      "post": {
        "tags": [
          "user"
        ],
        "summary": "Register Device",
        "description": "Register or update device for multi-device sync.",
        "operationId": "register_device_api_v1_user_device_post",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRegister"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/user/profile": {
      "get": {
        "tags": [
          "user"
        ],
        "summary": "Get Profile",
        "description": "Get user profile info.",
        "operationId": "get_profile_api_v1_user_profile_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DeviceRegister": {
        "properties": {
          "device_hash": {
            "type": "string",
            "title": "Device Hash"
          },
          "platform": {
            "type": "string",
            "title": "Platform"
          },
          "app_version": {
            "type": "string",
            "title": "App Version",
            "default": "0.1.0"
          }
        },
        "type": "object",
        "required": [
          "device_hash",
          "platform"
        ],
        "title": "DeviceRegister"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "PercentileResponse": {
        "properties": {
          "percentile": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Percentile"
          },
          "total_blocks": {
            "type": "integer",
            "title": "Total Blocks"
          },
          "bracket": {
            "type": "string",
            "title": "Bracket"
          }
        },
        "type": "object",
        "required": [
          "percentile",
          "total_blocks",
          "bracket"
        ],
        "title": "PercentileResponse"
      },
...
      "ProfileResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "email": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Email"
          },
          "tier": {
            "type": "string",
            "title": "Tier"
          },
          "devices": {
            "type": "integer",
            "title": "Devices"
          },
          "member_since": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Member Since"
          }
        },
        "type": "object",
        "required": [
          "id",
          "email",
          "tier",
          "devices",
          "member_since"
        ],
        "title": "ProfileResponse"
      },
...
      "ScoreSync": {
        "properties": {
          "score": {
            "type": "integer",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "score"
        ],
        "title": "ScoreSync"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WeeklyAggregateSubmit": {
        "properties": {
          "total_checks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Checks"
          },
          "total_blocks": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Blocks"
          },
          "total_warnings": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Warnings",
            "default": 0
          },
          "total_trackers": {
            "type": "integer",
            "minimum": 0.0,
            "title": "Total Trackers",
            "default": 0
          },
          "score": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 100.0,
                "minimum": 0.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          }
        },
        "type": "object",
        "required": [
          "total_checks",
          "total_blocks"
        ],
        "title": "WeeklyAggregateSubmit"
      },
...
    }
  }
}